requirements.txt
phidata
duckduckgo-search
